requirements.txt
alembic==1.16.4
fastapi==0.116.0
uvicorn==0.35.0
sqlalchemy==2.0.41
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pillow==11.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator==2.2.0
requests==2.32.4
PyMySQL==1.1.0
